Split extension implementation for rails 3/4
